Fix warnings

+++ MonsterChase/Engine/Vector2D.h
@@ -43,7 +43,7 @@ inline Vector2D operator +(const Vector2D &lhs, const Vector2D &rhs) {
 }
 
 inline bool operator ==(const Vector2D &lhs, const Vector2D &rhs) {
-	return is_approx_equals (lhs.x(), rhs.x(), 0.00001) && is_approx_equals(lhs.y(), rhs.y(), 0.00001);
+	return is_approx_equals (lhs.x(), rhs.x(), static_cast<float>(0.00001)) && is_approx_equals(lhs.y(), rhs.y(), static_cast<float>(0.00001));
 }
 
 inline bool operator !=(const Vector2D &lhs, const Vector2D &rhs) {
